finalise help

+++ core/src/mini73/TextBuffer.java
@@ -68,7 +68,7 @@ public class TextBuffer {
         updated = true;
     }
 
-    public float getSize() {
+    public int getSize() {
         return lines.size();
     }
